+++ lib/src/components/widgets/zpl_divider.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import '../base/zpl_component.dart';
 import '../../layout/geometry.dart';
 import '../../compiler/zpl_context.dart';
@@ -15,8 +16,21 @@ class ZplDivider extends ZplComponent {
   }
 
   @override
+  void finalizeLayout(ZplOffset absoluteOffset) {
+    setOffset(absoluteOffset);
+  }
+
+  @override
   void compile(ZplContext context) {
     context.addCommand('^FO${offset.dx.toInt()},${offset.dy.toInt()}');
     context.addCommand('^GB${size.width.toInt()},${thickness.toInt()},${thickness.toInt()}^FS\n');
+  }
+
+  @override
+  void paint(Canvas canvas, Offset offset) {
+    canvas.drawRect(
+      Rect.fromLTWH(this.offset.dx, this.offset.dy, size.width, thickness),
+      Paint()..color = Colors.black,
+    );
   }
 }